requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
auth==0.5.3
azure-core==1.29.4
azure-storage-blob==12.19.0
billiard==4.2.0
blinker==1.6.2
celery==5.3.5
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
dataclass-wizard==0.22.2
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.6
Django==4.2.1
django-autoslug==1.9.9
django-cors-headers==4.0.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-rest-passwordreset==1.3.0
django-storages==1.14.2
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.3
facebook-sdk==3.1.0
falcon==3.1.1
filelock==3.9.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
isodate==0.6.1
jwcrypto==1.5.1
kombu==5.3.3
MarkupSafe==2.1.3
mongoengine==0.27.0
msal==1.24.1
msal-extensions==1.0.0
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.25.2
numpy-quaternion==2023.0.2
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.0.0
portalocker==2.8.2
prompt-toolkit==3.0.41
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
PyMySQL==1.0.3
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-linkedin==4.1
python3-openid==3.2.0
pytz==2023.3
# pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
scipy==1.11.4
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
sqlparse==0.4.4
telegram==0.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
tweepy==4.14.0
twilio==8.5.0
twython==3.9.1
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.4
vine==5.1.0
virtualenv==20.19.0
wcwidth==0.2.10
webdriver-manager==4.0.0
Werkzeug==2.3.7
whitenoise==6.5.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
